+++ sol-did/programs/sol-did/src/lib.rs
@@ -11,15 +11,15 @@ unused_imports,
 // clippy::pedantic
 )]
 
+pub mod constants;
 pub mod integrations;
+pub mod legacy;
 pub mod state;
+pub mod utils;
 
-mod constants;
 mod errors;
 mod instructions;
-mod legacy;
 mod security_txt;
-mod utils;
 
 use crate::constants::DID_ACCOUNT_SEED;
 use crate::state::DidAccount;